forward/requirements.txt
backports.functools-lru-cache==1.5
certifi==2019.3.9
chardet==3.0.4
cheroot==6.5.4
CherryPy==18.1.1
Click==7.0
CoAPthon3==1.0.1
Flask==1.0.2
frozendict==1.2
idna==2.8
itsdangerous==1.1.0
jaraco.functools==2.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.0.0
netifaces==0.10.9
paho-mqtt==1.4.0
portend==2.4
pymongo==3.7.2
python3-ngrok==0.2
pytz==2018.9
# pywin32==224
redis==3.2.1
requests==2.21.0
schema==0.6.7
six==1.12.0
tempora==1.14
urllib3==1.24.1
Werkzeug==0.15.2
zc.lockfile==1.4
